* 'develop' of github.com:lesstif/jira-rest-client:
  Problems related with the field "workratio"
  add new project property(projectTypeKey)

+++ src/main/java/com/lesstif/jira/issue/IssueFields.java
@@ -72,7 +72,7 @@ public class IssueFields {
 	
 	private String[] labels;
 	
-	private Integer workratio;
+	private Long workratio;
 	
 	private String environment;
